Revert "New Row and extends by RowResult"
This reverts commit 5d14dfc2
