+++ src/Db/Row.php
@@ -9,15 +9,28 @@ namespace Forrest79\PhPgSql\Db;
 class Row implements \ArrayAccess, \IteratorAggregate, \Countable, \JsonSerializable
 {
 	/** @var array<string, string|NULL> */
-	protected $rawValues;
+	private $rawValues;
+
+	/** @var array<string, string> */
+	private $columnsDataTypes;
+
+	/** @var DataTypeParser */
+	private $dataTypeParser;
 
 	/** @var array<string, mixed> */
-	protected $values;
+	private $values;
 
 
-	public function __construct(array $values)
+	/**
+	 * @param array<string, mixed> $values
+	 * @param array<string, string> $columnsDataTypes
+	 */
+	public function __construct(array $values, array $columnsDataTypes, DataTypeParser $dataTypeParser)
 	{
 		$this->rawValues = $values;
+		$this->columnsDataTypes = $columnsDataTypes;
+		$this->dataTypeParser = $dataTypeParser;
+
 		$this->values = \array_fill_keys(\array_keys($values), NULL);
 	}
 
@@ -58,6 +71,9 @@ class Row implements \ArrayAccess, \IteratorAggregate, \Countable, \JsonSerializ
 	 */
 	public function toArray(): array
 	{
+		foreach ($this->rawValues as $key => $value) { // intentionally not using array_keys($this->rawValues) as $key - this is 2x faster
+			$this->parseValue($key);
+		}
 		return $this->values;
 	}
 
@@ -139,9 +155,24 @@ class Row implements \ArrayAccess, \IteratorAggregate, \Countable, \JsonSerializ
 	 * @return mixed
 	 * @throws Exceptions\RowException
 	 */
-	protected function getValue(string $key)
+	private function getValue(string $key)
 	{
+		if (!\array_key_exists($key, $this->values)) {
+			throw Exceptions\RowException::noParam($key);
+		}
+
+		if (\array_key_exists($key, $this->rawValues)) {
+			$this->parseValue($key);
+		}
+
 		return $this->values[$key];
+	}
+
+
+	private function parseValue(string $key): void
+	{
+		$this->values[$key] = $this->dataTypeParser->parse($this->columnsDataTypes[$key], $this->rawValues[$key]);
+		unset($this->rawValues[$key]);
 	}
